pyproject.toml
[project]
name = "secretbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cookiecutter>=2.7.1",
    "django>=6.0.3",
    "django-browser-reload>=1.21.0",
    "django-crispy-forms>=2.6",
    "django-debug-toolbar>=6.3.0",
    "django-environ>=0.13.0",
    "django-extensions>=4.1",
    "django-stubs-ext>=6.0.2",
    "django-tailwind>=4.4.2",
    "gitingest>=0.3.1",
    "gunicorn>=25.3.0",
    "pillow>=12.2.0",
]
